+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.16.1-1] - 2024-09-24
+### Changed
+- Improve error messages in user import
+  - The error messages now look more like the error messages in single user creation
+  - There are specific error messages for unique mail, username too long/short, username with invalid characters
+
 ## [v1.16.0-1] - 2024-09-19
 ### Changed
 - Relicense to AGPL-3.0-only